requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.20
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
torch==2.5.0
torchvision==0.20.0
ultralytics==8.3.57
boto3==1.34.162
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
